backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-multipart
Jinja2
pytesseract
pdf2image
pillow
sentence-transformers
numpy
